Adding sorting. #broken #wip
There’s a small issue, not breaking but still.
After sorting PATCH to the todo controller, the redirect to the project is done to another PATCH … returning 400 since no project params are being sent. This is done in other calls to re-render the Turbolink sections in the project correctly as GET.